Refactor gold-related models and API documentation
- Updated `HomeGoldScreen.kt` and `GoldListItem.kt` to use Double for carat, grams, and price instead of BigDecimal and Int.
- Added new `DtoGoldTax` model to represent gold tax data.
- Modified API documentation to reflect changes in data types for price and tax-related fields.
- Changed base URL in API client and documentation from `localhost:3000` to `localhost:8080`.
- Updated various model classes to use Double for price and tax rate fields.
- Adjusted Swagger and OpenAPI documentation to align with the new data structure and types.

+++ android/app/src/main/java/id/my/rizalanggoro/arta/feature/home/presentation/dashboard/gold/component/LatestGold.kt
@@ -80,19 +80,19 @@ private fun Preview() {
         golds = List(3) {
             DtoGold(
                 data = DomainGold(
-                    carat = 24.toBigDecimal(),
+                    carat = 24.0,
                     createdAt = "2026-05-25T14:38:00.000+07:00",
                     date = "2026-05-25T14:38:00.000+07:00",
-                    grams = 3.3.toBigDecimal(),
+                    grams = 3.3,
                     id = 1,
                     notes = "",
-                    price = 1500000,
+                    price = 1500000.0,
                     type = "jewelry",
                     updatedAt = "2026-05-25T14:38:00.000+07:00",
                     walletId = 1
                 ),
-                profit = ((it - 1) * 500000).toBigDecimal(),
-                sellPrice = (1500000 + ((it - 1) * 500000)).toBigDecimal(),
+                profit = ((it - 1) * 500000.0),
+                sellPrice = (1500000 + ((it - 1) * 500000.0)),
             )
         }
     )
