+++ android/app/src/main/java/id/my/rizalanggoro/arta/feature/home/presentation/gold/HomeGoldScreen.kt
@@ -17,10 +17,10 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
-import id.my.rizalanggoro.arta.core.utils.LocalBackStack
 import id.my.rizalanggoro.arta.core.application.Routes
 import id.my.rizalanggoro.arta.core.event.AppEvent
 import id.my.rizalanggoro.arta.core.event.AppEventBus
+import id.my.rizalanggoro.arta.core.utils.LocalBackStack
 import id.my.rizalanggoro.arta.openapi.models.DomainGold
 import id.my.rizalanggoro.arta.openapi.models.DtoGold
 import id.my.rizalanggoro.arta.shared.component.EmptyPlaceholder
@@ -110,19 +110,19 @@ private fun Preview() {
                 golds = List(5) {
                     DtoGold(
                         data = DomainGold(
-                            carat = 24.toBigDecimal(),
+                            carat = 24.0,
                             createdAt = "2026-05-25T14:38:00.000+07:00",
                             date = "2026-05-25T14:38:00.000+07:00",
-                            grams = 3.3.toBigDecimal(),
+                            grams = 3.3,
                             id = 1,
                             notes = "",
-                            price = 1500000,
+                            price = 1500000.0,
                             type = "jewelry",
                             updatedAt = "2026-05-25T14:38:00.000+07:00",
                             walletId = 1
                         ),
-                        profit = ((it - 1) * 500000).toBigDecimal(),
-                        sellPrice = (1500000 + ((it - 1) * 500000)).toBigDecimal(),
+                        profit = ((it - 1) * 500000.0),
+                        sellPrice = (1500000 + ((it - 1) * 500000.0)),
                     )
                 },
                 isLoading = false,
